tests/requirements.txt
# Testing Requirements for DexAgent
# Install with: pip install -r tests/requirements.txt

# Core testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0
pytest-xdist>=3.3.0  # For parallel test execution
pytest-html>=3.2.0   # For HTML test reports
pytest-cov>=4.1.0    # For coverage reports
pytest-mock>=3.11.0  # For mocking

# HTTP and API testing
requests>=2.31.0
httpx>=0.24.0  # For async HTTP requests
aiohttp>=3.8.0

# WebSocket testing
websockets>=11.0.0
python-socketio[client]>=5.8.0

# Database testing
psycopg2-binary>=2.9.7
SQLAlchemy>=2.0.0
alembic>=1.11.0

# Browser automation (E2E testing)
playwright>=1.37.0
selenium>=4.11.0

# Performance and load testing
locust>=2.16.0  # For load testing
psutil>=5.9.0   # For system resource monitoring
memory-profiler>=0.61.0  # For memory profiling

# Data manipulation and analysis
pandas>=2.0.0  # For test data analysis
numpy>=1.24.0  # For numerical operations

# Utilities
faker>=19.3.0   # For generating test data
factory-boy>=3.3.0  # For test object factories
freezegun>=1.2.0  # For time mocking
responses>=0.23.0  # For HTTP response mocking

# Monitoring and reporting
pytest-benchmark>=4.0.0  # For performance benchmarking
pytest-profiling>=1.7.0  # For profiling
pytest-monitor>=1.6.0    # For resource monitoring

# Code quality
flake8>=6.0.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.1.0
PyYAML>=6.0

# Async testing support
anyio>=3.7.0
trio>=0.22.0

# Network and protocol testing
scapy>=2.5.0  # For network packet analysis (optional)
paramiko>=3.2.0  # For SSH testing (optional)

# Image and media testing (for screenshots, etc.)
Pillow>=10.0.0
imageio>=2.31.0

# Concurrency testing
concurrent-futures>=3.1.1
threading2>=0.1.0

# JSON and data validation
jsonschema>=4.19.0
cerberus>=1.3.4

# Time and date utilities
python-dateutil>=2.8.0
arrow>=1.2.0

# Logging and debugging
loguru>=0.7.0
rich>=13.5.0  # For rich console output

# CLI testing
click>=8.1.0
typer>=0.9.0

# File and path utilities
pathlib2>=2.3.7
watchdog>=3.0.0

# Encryption and security testing
cryptography>=41.0.0
bcrypt>=4.0.0

# Optional: GUI testing (if needed for agent GUI)
pyautogui>=0.9.54
pynput>=1.7.6

# Optional: Docker integration testing
docker>=6.1.0
testcontainers>=3.7.0

# Optional: Cloud testing
boto3>=1.28.0  # For AWS testing
google-cloud-storage>=2.10.0  # For GCP testing

# Development and debugging
ipdb>=0.13.0
pdbpp>=0.10.0
pytest-pdb>=0.2.0